CurveratureControl working!  Refactored to float for splines for speed.
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@2127 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/virtualslotcar/CurvatureBasedController.java
@@ -97,8 +97,8 @@ This still requires us to have an estimated relation between throttle and result
             currentTrackPos= newState.segmentIdx;
             // compute the curvature at throttleDelayMs in the future, given our measured speed
 
-            double timeStep = getThrottleDelayMs();
-            UpcomingCurvature curvature=track.getCurvature(currentTrackPos, 2, timeStep/1000.0, measuredSpeedPPS);
+            float timeStep = getThrottleDelayMs();
+            UpcomingCurvature curvature=track.getCurvature(currentTrackPos, 2, timeStep/1000, measuredSpeedPPS);
             // The first entry of the curvature is always the curvature at the current position
             // If you need the curvature one timestep ahead, you need to extract two points
             // and use the second entry "curvature.getCurvature(1)"
